Decouple backend from frontend
